Module Tezos_store.Consistency

val check_consistency : [ `Chain_dir ] Naming.directory -> Tezos_base__TzPervasives.Genesis.t -> unit Tezos_base__TzPervasives.tzresult Lwt.t

check_consistency store_dir genesis checks that the store is in a consistent state.

val fix_consistency : [ `Chain_dir ] Naming.directory -> Tezos_context.Context.index -> Tezos_base__TzPervasives.Genesis.t -> unit Tezos_base__TzPervasives.tzresult Lwt.t

fix_consistency store_dir context_index ~genesis aims to fix a store if it is in an inconsistent state. To do so, it will inspect the stored data and try infer all the missing or corrupted parts.