Module Tezos_store.Floating_block_index

Floating blocks index map

Block info stored in the index map

module Block_info : sig ... end

Module Block_info represents the values stored in the index which are used to locate the position of a block in a floating block store and its predecessors.

Key/value index associated to a floating block store where the key is a Block_hash.t and the value is Block_info.t.

include Index.S with type key = Tezos_base__TzPervasives.Block_hash.t and type value = Block_info.t

type t

type key = Tezos_crypto.Block_hash.t

type value = Block_info.t

type cache

val empty_cache : unit -> cache

val v : ?flush_callback:(unit -> unit) -> ?cache:cache -> ?fresh:bool -> ?readonly:bool -> ?throttle:[ `Block_writes | `Overcommit_memory ] -> log_size:int -> string -> t

val clear : t -> unit

val find : t -> key -> value

val mem : t -> key -> bool

val replace : ?overcommit:bool -> t -> key -> value -> unit

val filter : t -> ((key * value) -> bool) -> unit

val iter : (key -> value -> unit) -> t -> unit

val flush : ?no_callback:unit -> ?with_fsync:bool -> t -> unit

val close : ?immediately:unit -> t -> unit

val sync : t -> unit

val is_merging : t -> bool

val merge : t -> unit

val try_merge : t -> unit

module Checks : sig ... end