Module Store.Chain

The module for handling chain-related operations such as setting the head of the chain, updating the chain state, forking testchains, ...

type nonrec chain_store = chain_store

The abstract type alias of a chain store.

type t = chain_store

The type alias of chain_store.

type testchain

The abstract type for testchain.

val global_store : chain_store -> store

global_store chain_store returns the global store of chain_store allowing to retrieve global infos.

val chain_id : chain_store -> Tezos_crypto.Chain_id.t

chain_id chain_store returns chain id of chain_store.

val chain_dir : chain_store -> [ `Chain_dir ] Naming.directory

chain_dir chain_store returns the path of directory of chain_store.

val history_mode : chain_store -> Tezos_shell_services.History_mode.t

history_mode chain_store returns the history mode of the chain_store.

val genesis : chain_store -> Tezos_base__TzPervasives.Genesis.t

genesis chain_store returns the Genesis.t of the chain_store.

val genesis_block : chain_store -> Block.t Lwt.t

genesis chain_store returns the genesis block of the chain_store.

val current_head : chain_store -> Block.t Lwt.t

current_head chain_store returns the current head of the chain_store.

val expiration : chain_store -> Tezos_base__TzPervasives.Time.Protocol.t option

expiration chain_store returns the expiration date of the testchain's chain_store.

val checkpoint : chain_store -> Store_types.block_descriptor Lwt.t

checkpoint chain_store returns the checkpoint associated to the chain_store.

The checkpoint is a block descriptor (Store_types.block_descriptor) pointing to a block that must be part of the chain. The checkpoint maintains the set of following invariants:

The store will only accept blocks that are above its checkpoint's level.

The checkpoint is updated periodically such that the following invariant holds: checkpoint.level >= all_head.last_allowed_fork_level

The checkpoint will tend to designate the highest block among all chain head's last_allowed_fork_level in a normal mode. This is not always true. i.e. after a snapshot import where the checkpoint will be set as the imported block and when the target block is reached, the checkpoint will be set at this point.

val target : chain_store -> Store_types.block_descriptor option Lwt.t

target chain_store returns the target block associated to the chain_store if there is one or None if it has already been reached.

The target is a (Store_types.block_descriptor) pointing to a future block that must be reached by the chain. When a block is received, if its level is equal to the target's block level then its hash must be also be the same. Otherwise, the block is considered as invalid and will be discarded.

The target should only be set manually. Whenever the target block is reached, the checkpoint will be updated to this block and the target will be set to None.

val savepoint : chain_store -> Store_types.block_descriptor Lwt.t

savepoint chain_store returns the savepoint associated to the chain_store.

The savepoint is a block descriptor (Store_types.block_descriptor) pointing to the lowest level block that has its metadata and/or context pruned with the following invariant:

is_stored(block) \and has_metadata(block) => block.level >= savepoint.level

For Full and Rolling history modes, the savepoint will be periodically updated at each store merge which happens when:

pred(head).last_allowed_fork_level < head.last_allowed_fork_level

On Archive history mode: savepoint = genesis.

val caboose : chain_store -> Store_types.block_descriptor Lwt.t

caboose chain_store returns the caboose associated to the chain_store.

The caboose is a block descriptor (Store_types.block_descriptor) pointing to the lowest level block that we know present in the chain store, if the history mode is not Archive, it might not contain its metadata depending on the current savepoint:

is_stored(block) => block.level >= caboose.level

On Archive and Full history mode: caboose = genesis

val mempool : chain_store -> Tezos_base__TzPervasives.Mempool.t Lwt.t

mempool chain_store returns the mempool associated to the chain_store.

val set_mempool : chain_store -> head:Tezos_crypto.Block_hash.t -> Tezos_base__TzPervasives.Mempool.t -> unit Tezos_base__TzPervasives.tzresult Lwt.t

set_mempool chain_store ~head mempool sets the mempool of the chain_store. Does nothing if head is not current_head which might happen when a new head concurrently arrives just before this operation being called.

val live_blocks : chain_store -> (Tezos_crypto.Block_hash.Set.t * Tezos_crypto.Operation_hash.Set.t) Lwt.t

live_blocks chain_store returns the set of previously computed live blocks for the current_head's chain_store.

val compute_live_blocks : chain_store -> block:Block.t -> (Tezos_crypto.Block_hash.Set.t * Tezos_crypto.Operation_hash.Set.t) Tezos_base__TzPervasives.tzresult Lwt.t

compute_live_blocks ~block chain_store computes the set of live blocks and live operations relative to block. Does nothing if block is the chain_store's current head as it was previously updated in set_head.

Note: this operation should not be costly in most cases as recent blocks and operations are expected to be in a cache.

val set_head : chain_store -> Block.t -> Block.t option Tezos_base__TzPervasives.tzresult Lwt.t

set_head chain_store block promotes the block as head of the chain_store and triggers an asynchronous store merge if a cycle is ready to be cemented. Triggering a merge will update the savepoint, checkpoint and caboose consistently with the chain_store's history mode. This function returns the previous head or None if the given block is below one of the current known heads. If block belongs to a new branch, the previous head will also be stored as an alternate head. Setting a new head will fail when the block is not fit to be promoted as head: too old or no metadata.

After a merge:

The checkpoint is updated to lafl(new_head) if it was below this level or unchanged otherwise;

The savepoint will be updated to : min(max_op_ttl(lafl(new_head)), lafl(new_head) - <cycle_length> * <history_mode_offset>) or will remain 0 in Archive mode;

The caboose will be updated to the same value as the savepoint in Rolling mode.

Note: lafl(new_head) is the last allowed fork level of the new head.

Warnings:

We expect blocks to be sequentially promoted as head using this function;

If a merge is triggered while another is happening, this function will block until the first merge is resolved.

val known_heads : chain_store -> Store_types.block_descriptor list Lwt.t

known_heads chain_store returns the list of alternate heads for chain_store.

val is_ancestor : chain_store -> head:Store_types.block_descriptor -> ancestor:Store_types.block_descriptor -> bool Lwt.t

is_ancestor chain_store ~head ~ancestor checks whether the ancestor is a predecessor or head in chain_store.

val is_in_chain : chain_store -> Store_types.block_descriptor -> bool Lwt.t

is_in_chain chain_store block_descr checks that block_descr is an ancestor of chain_store's current head.

val is_acceptable_block : chain_store -> Store_types.block_descriptor -> bool Lwt.t

is_acceptable_block chain_store block_descr checks if block_descr would be a valid block to be stored in chain_store. Its predecessor is supposed to be already stored.

val best_known_head_for_checkpoint : chain_store -> checkpoint:Store_types.block_descriptor -> Block.t Tezos_base__TzPervasives.tzresult Lwt.t

best_known_head_for_checkpoint chain_store checkpoint returns the fittest block among known heads in chain_store to be promoted as checkpoint if there is one.

val compute_locator : chain_store -> ?⁠max_size:int -> Block.t -> Tezos_base__TzPervasives.Block_locator.seed -> Tezos_base__TzPervasives.Block_locator.t Lwt.t

compute_locator chain ?max_size head seed computes a locator of the chain from head to the chain's caboose or until the locator contains max_size steps. max_size defaults to 200.

val compute_protocol_locator : chain_store -> ?⁠max_size:int -> proto_level:int -> Tezos_base__TzPervasives.Block_locator.seed -> Tezos_base__TzPervasives.Block_locator.t option Lwt.t

compute_protocol_locator chain ?max_size ~proto_level seed computes a locator for a specific protocol of level proto_level in the chain from the latest block with this protocol to its activation block or until the locator contains max_size steps. max_size defaults to 200.

val set_target : chain_store -> Store_types.block_descriptor -> unit Tezos_base__TzPervasives.tzresult Lwt.t

set_target chain_store target_descr sets the target for chain_store. If target_descr is already known, set_checkpoint will be internally called. Fails if the target is below the current checkpoint.

val testchain : chain_store -> testchain option Lwt.t

testchain chain_store returns the active testchain to the given chain_store if it has been instantiated.

val testchain_forked_block : testchain -> Tezos_crypto.Block_hash.t

testchain_forked_block testchain returns the hash of the forked block of its parent chain.

val testchain_store : testchain -> chain_store

testchain_store testchain returns the chain store associated to this testchain.

val fork_testchain : chain_store -> testchain_id:Tezos_crypto.Chain_id.t -> forked_block:Block.t -> genesis_hash:Tezos_crypto.Block_hash.t -> genesis_header:Tezos_base__TzPervasives.Block_header.t -> test_protocol:Tezos_crypto.Protocol_hash.t -> expiration:Tezos_base__TzPervasives.Time.Protocol.t -> testchain Tezos_base__TzPervasives.tzresult Lwt.t

fork testchain chain_store ~testchain_id ~forked_block ~genesis_hash ~genesis_header ~test_protocol ~expiration forks a testchain and activates it for chain_store. If a testchain with testchain_id already existed, it is then loaded from the store. If it was already activated, does nothing. It also registers this chain in the set of activated testchains.

val shutdown_testchain : chain_store -> unit Tezos_base__TzPervasives.tzresult Lwt.t

shutdown_testchain chain_store closes and deactivates the testchain running for chain_store. Does nothing if no testchain is found running for chain_store.

Chain's protocols

val find_activation_block : chain_store -> protocol_level:int -> Store_types.Protocol_levels.activation_block option Lwt.t

find_activation_block chain_store ~protocol_level returns the block that activated the protocol of level protocol_level.

val find_protocol : chain_store -> protocol_level:int -> Tezos_crypto.Protocol_hash.t option Lwt.t

find_protocol chain_store ~protocol_level returns the protocol with the level protocol_level.

val all_protocol_levels : chain_store -> Store_types.Protocol_levels.activation_block Store_types.Protocol_levels.t Lwt.t

all_protocol_levels chain_store returns all the protocols registered in chain_store.

val may_update_protocol_level : chain_store -> protocol_level:int -> (Block.block * Tezos_crypto.Protocol_hash.t) -> unit Tezos_base__TzPervasives.tzresult Lwt.t

may_update_protocol_level chain_store ~protocol_level (block, ph) updates the protocol level for the protocol ph in chain_store with the activation block. If a previous entry is found, does nothing.

val watcher : chain_store -> Block.t Lwt_stream.t * Lwt_watcher.stopper

watcher chain_store instantiates a new block watcher for chain_store.

val get_rpc_directory : chain_store -> Block.t -> (chain_store * Block.t) Tezos_rpc.RPC_directory.t option Lwt.t

get_rpc_directory chain_store block returns the RPC directory associated to the block.

val set_rpc_directory : chain_store -> Tezos_crypto.Protocol_hash.t -> (chain_store * Block.t) Tezos_rpc.RPC_directory.t -> unit Lwt.t

set_rpc_directory chain_store ph rpc_directory sets a rpc_directory for the protocol ph in chain_store.