Module Tezos_store.Store_types

Global types used in the store library

type block_descriptor = Tezos_crypto.Block_hash.t * int32

The type used to describe a block pointer i.e. its hash and level.

val block_descriptor_encoding : block_descriptor Tezos_base__TzPervasives.Data_encoding.t

Encoding for block_descriptor.

type chain_config = {

history_mode : Tezos_shell_services.History_mode.t;

genesis : Tezos_base.Genesis.t;

expiration : Tezos_base.Time.Protocol.t option;

}

The type for chain configuration.

val chain_config_encoding : chain_config Tezos_base__TzPervasives.Data_encoding.t

Encoding for chain_config.

val pp_block_descriptor : Stdlib.Format.formatter -> block_descriptor -> unit

Pretty-printer for block_descriptor

type invalid_block = {

level : int32;

errors : Tezos_error_monad.TzCore.error list;

}

The type used to store an invalid block's value. We only retain the level and the errors encountered during validation. These values should be indexed by the block's hash.

val invalid_block_encoding : invalid_block Tezos_base__TzPervasives.Data_encoding.t

Encoding for invalid_block.

module Protocol_levels : sig ... end

Module Protocol_levels represents an associative map of protocol levels to corresponding blocks which supposedly activate new protocols, that is to say blocks that acknowledge a protocol change in the next block.