Module Tezos_store__Legacy

Legacy storage upgrade

The legacy store upgrade aims to migrate a storage using the LMDB backend (v0.0.4) to the new store representation (v0.0.5). This migration is available for any store running v0.0.4 with any history mode. The upgrade procedure is going to retrieve each the block and its associated data available in the old store, convert and store them in the new backend. It will preserve all the information originally contained in the LMDB store such as the current head, checkpoint, savepoint and caboose.