Module Tezos_store__Legacy.Hardcoded

Module for handling values needed by the legacy upgrade. It exposes the cycle_length to allow upgrading supported_networks only. These values are mandatory as the new backend regroup blocks by cycles, something which was not available in the previous representation when the metadata of blocks is not fully available.

type network = {

name : Tezos_base__TzPervasives.Distributed_db_version.Name.t;

cycle_length : int;

}

val supported_networks : network list

val cycle_length : chain_name:Tezos_base__TzPervasives.Distributed_db_version.Name.t -> int

val check_network : chain_name:Tezos_base__TzPervasives.Distributed_db_version.Name.t -> unit Tezos_base__TzPervasives.tzresult Lwt.t